Preseve load balancer between configuration reloads
Related issue: https://github.com/traefik/traefik/issues/8275

I consider this very much just a hack.  I'm not sure if has a lot of
performance downsides.

Traefik currently creates a new copy of the configuration and a whole new http Handler every time
kubernetes or any provider detects a change. Then it swaps the http
Handler out for the new one.

Down side to that is it trashes the state for roundrobin.  Causing every
configuration change back to the beginning of the pool. Stressing the
servers at the front part of the pool out more if you have a lot of
configuration reloads.

My solution is just to create a state machine for the loadbalancers and
re-use from that state.  This did cause me to have to fork the
vulcan/oxy roundrobin package because every add or remove they trash the
index counter which would defeat the entire purpose.

So in short.. doing this just to get some comments. Maybe use as a work
around if we determine the issue warrents a quick patch